move redis related files to redis folder
