+++ packages/indexer/src/redis/rangeQueryStore.ts
